refactor: move IO functions used by the commandline to separate file, add auspice_json export
